+++ lib/solvers/PortPointPathingSolver/tinyhypergraph/TinyHypergraphPortPointPathingSolver.ts
@@ -23,8 +23,9 @@ import {
   type TinyHyperGraphSectionSolverOptions,
   type TinyHyperGraphSolverOptions,
 } from "tiny-hypergraph/lib/index"
-import { checkIfConnectionPointIsInRegion } from "../hgportpointpathingsolver/checkIfConnectionPointIsInRegion"
 import type { HgPortPointPathingSolverParams } from "../hgportpointpathingsolver/types"
+import { createTinyRouteNetIndexer } from "./createTinyRouteNetIndexer"
+import { getRegionNetIdByRegionId } from "./getRegionNetIdByRegionId"
 
 type RouteMetadata = {
   connectionId: string
@@ -263,53 +264,11 @@ const toSerializedPortData = (
 const buildSerializedTinyGraph = (
   params: HgPortPointPathingSolverParams,
 ): SerializedHyperGraph => {
-  const netIndexById = new Map<string, number>()
-  const getNetIndex = (routeMetadata: RouteMetadata) => {
-    const netId =
-      routeMetadata.mutuallyConnectedNetworkId ?? routeMetadata.connectionId
-    let netIndex = netIndexById.get(netId)
-    if (netIndex === undefined) {
-      netIndex = netIndexById.size
-      netIndexById.set(netId, netIndex)
-    }
-    return netIndex
-  }
-  const regionNetCandidates = new Map<string, Set<number>>()
-  for (const connection of params.connections) {
-    const routeMetadata: RouteMetadata = {
-      connectionId: connection.connectionId,
-      mutuallyConnectedNetworkId:
-        connection.mutuallyConnectedNetworkId ?? connection.connectionId,
-      simpleRouteConnection: connection.simpleRouteConnection,
-    }
-    const routeNetIndex = getNetIndex(routeMetadata)
-    for (const point of connection.simpleRouteConnection?.pointsToConnect ??
-      []) {
-      for (const region of params.graph.regions) {
-        if (
-          !checkIfConnectionPointIsInRegion({
-            point,
-            region,
-            layerCount: params.layerCount,
-          })
-        ) {
-          continue
-        }
-
-        let netCandidates = regionNetCandidates.get(region.regionId)
-        if (!netCandidates) {
-          netCandidates = new Set<number>()
-          regionNetCandidates.set(region.regionId, netCandidates)
-        }
-        netCandidates.add(routeNetIndex)
-      }
-    }
-  }
-  const regionNetIdByRegionId = new Map<string, number>()
-  for (const [regionId, netCandidates] of regionNetCandidates) {
-    if (netCandidates.size !== 1) continue
-    regionNetIdByRegionId.set(regionId, [...netCandidates][0]!)
-  }
+  const getNetIndex = createTinyRouteNetIndexer()
+  const regionNetIdByRegionId = getRegionNetIdByRegionId({
+    params,
+    getNetIndex,
+  })
 
   const regions: SerializedHyperGraph["regions"] = params.graph.regions.map(
     (region) => ({
