+++ lib/solvers/PortPointPathingSolver/tinyhypergraph/createTinyRouteNetIndexer.ts
@@ -1,0 +1,23 @@
+export type TinyRouteNetIndexSource = {
+  connectionId: string
+  mutuallyConnectedNetworkId?: string
+}
+
+export type TinyRouteNetIndexer = (
+  routeMetadata: TinyRouteNetIndexSource,
+) => number
+
+export function createTinyRouteNetIndexer(): TinyRouteNetIndexer {
+  const netIndexById = new Map<string, number>()
+
+  return (routeMetadata: TinyRouteNetIndexSource): number => {
+    const netId =
+      routeMetadata.mutuallyConnectedNetworkId ?? routeMetadata.connectionId
+    let netIndex = netIndexById.get(netId)
+    if (netIndex === undefined) {
+      netIndex = netIndexById.size
+      netIndexById.set(netId, netIndex)
+    }
+    return netIndex
+  }
+}
